[less] #390 - Add support for "package!" scheme for resolving dependencies/imports in the same package but in a different Jar file
