+++ bootstrap-less/src/test/java/de/agilecoders/wicket/less/BootstrapLessTest.java
@@ -9,6 +9,7 @@ import org.junit.Test;
 
 import java.net.URL;
 
+import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
@@ -49,7 +50,6 @@ public class BootstrapLessTest {
         LessSource.URLSource lessSource = less.getLessSource(res, null);
         String css = less.getCss(lessSource);
         assertThat(css, is(equalTo(".rule {\n  background: #428bca;\n}\n")));
-
     }
 
     @Test
@@ -60,17 +60,17 @@ public class BootstrapLessTest {
         LessSource.URLSource lessSource = less.getLessSource(res, null);
         String css = less.getCss(lessSource);
         assertThat(css, is(equalTo(".classPathImported {\n  color: #333;\n}\n")));
-
     }
 
     @Test
     public void importPackage() throws Exception {
         LessCacheManager less = LessCacheManager.get();
-        URL res = BootstrapLessTest.class.getResource("1.less");
+        URL res = BootstrapLessTest.class.getResource("package-dependency-1.less");
 
         LessSource.URLSource lessSource = less.getLessSource(res, BootstrapLessTest.class.getName());
         String css = less.getCss(lessSource);
-        assertThat(css, is(equalTo("body {\n  background-color: #808080;\n}\n")));
-
+        assertThat(css, containsString("package1"));
+        assertThat(css, containsString("package2"));
+        assertThat(css, containsString("package3"));
     }
 }
