--- bootstrap-less/src/test/resources/de/agilecoders/wicket/less/1.less
@@ -1,1 +1,0 @@
-@import "package!2.less";
