--- bootstrap-less/src/test/resources/de/agilecoders/wicket/less/2.less
@@ -1,3 +1,0 @@
-body {
-  background-color: #808080;
-}
