+++ bootstrap-less/src/test/resources/de/agilecoders/wicket/less/package-dependency-1.less
@@ -1,0 +1,2 @@
+package1 {name: "package1";}
+@import "package!package-dependency-2.less";
