+++ bootstrap-less/src/test/resources/de/agilecoders/wicket/less/package-dependency-2.less
@@ -1,0 +1,2 @@
+package2 {name: "package2";}
+@import "package!test/package-dependency-3.less";
