+++ bootstrap-less/src/test/resources/de/agilecoders/wicket/less/test/package-dependency-3.less
@@ -1,0 +1,1 @@
+package3 {name: "package3";}
